requirements.txt
asgiref==3.5.2
Django==3.2.13
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
supervisor==4.2.4
Sphinx==5.0.1
flake8==4.0.1
coverage==6.4.1
twine==4.0.1
whitenoise==6.2.0
django-on-heroku==1.1.2
whatever==0.6